From SVN: Fixed size allocation bug in MapView
The TileSet GConf code introduced this bug.  Start Questar (at the
normal zoom) and make the window as small as it'll go.  Then zoom in all
the way and quit.  Now when you restart Questar the window will be way
bigger.  This is because MapView figures it's size allocation from the
width and height of a Tile in the TileSet.  Now we just set the size
allocation to 640x480.