Bug Fix: MapView incorrectly redraws when an Actor moves
This bug shows up when the hero is not moving but takes a turn by
pressing '5' on the keypad. When an actor moves it's old location and
new location are invalidated, so they can be redrawn. The invalidation
area was being set wrong, so part of an actor would be left behind and
part would not be drawn at the new location. I missed this bug, because
when I was testing the hero centering code, I was moving around the hero
alot.