Users of ConfigurationClient.SyncToggleAction() don't compile anymore
Recent versions of mcs and friends have a regression that prevents users
of ConfigurationClient.SyncToggleAction() (MainWindow and WorldView)
from compiling. I'm not exactly sure why, but the compiler crashes, so
this is a problem with them.

This means fullscreen mode no longer works and showing grid lines on the
world view no longer works.

This will be fixed when the compiler works again.