HitPoints: Cleanups and fixes
Added file header.

Got rid of unused HitPointsEventArgs constructor.

Fixed HitPoints.Changed event raising. The new value wasn't being set
before the event was fired. This made the handler method's
args.OldHitPoints to be the same and sender.

New property, HitPoints.AsDouble, returns the HitPoints object as a
double. A HitPoints object of 100/400 would be returned as 0.25.

Fixed HitPoints equality. Comparing to null caused an Exception to be
thrown. The new code is similar to Location's equality implementation.