HitPointsWidget: Minor cleanups
New static class, TangoColors, contains public static Color fields of
the Tango color schema.

HitPointsWidget uses the colors defined in TangoColors.

HitPointsWidget uses ClosePath() instead of drawing a line to the end.
This makes the drawing not look so weird.

The HitPointsWidget's stroke is now 4 pixels. And it's radius is a
little less to make up for it.

New method, ColorHelper.FromString(), to create a Color from a string.
It uses Gdk.Color.Parse() internally, so any string that works for it
will work here.