Added DoubleAnimation object to animate double values
DoubleAnimation will soon be abstracted into an interface and abstract
base class. Currently, allows specifying duration and start and end
values. Later on, we'll allow multiple "stops" or "positions". NewFrame
event is fired on every new frame, so clients can update themselves.

HitPointsWidget uses this to animate when the HP changes.

This is currently buggy, but works alright.