Animations are now generic
DoubleAnimation has been replaced with a generic Animation<T> class.
It's now up to Animation<T>'s users to "transform" the animation from
one frame to the next. Animation<T> exposes more properties so users can
do this easily.

HitPointsChart uses an Animation<int> instead of DoubleAnimation. It now
uses regular degrees as it's angle and it converts that to radians
suitable for Cairo when drawing. This makes it easier to debug.