New Location methods, ActorsInRadius() and LocationsInRadius()
These new methods return a list of Actors or Locations in a given
radius. Right now they actually return Actors or Locations in a square
centered around the Location they were called on. Later when I figure
out circles/ellipses I'll fix it.

New static class LocationFactory. When attempting to create a
MapLocation for an invalid Map or Point, when need to be able to return
a NullLocation, this takes care of that.