Actors are now aware of other Actors
Each time an Actor moves it checks a radius around its location and
creates a list of visible Actors.

This code will eventually raise events (SightedActor, LostSightOfActor)
for each Actor that is newly sighted (not previously sighted, unless it
was lost sight of first) or sight is lost. It uses
Location.ActorsInRadius() which returns the Actors in a rectangle around
the Location, but will eventually use radius properly and use LOS to
determine sight.

This code will probably be put in a TakeTurn() method, so it can run
each turn, because it won't run if the Actor doesn't do a MoveAction and
this stuff should be updated everyturn.