Actor: Replaced Action property with TakeTurn() method
The Action property isn't actually gone, its now a protected read/write
property of AbstractActor, so Hero and Monster can change it.

The code for TakeTurn() in AbstractActor is virtual, so it can be
overridden and only executes the Action property.

Monster and Hero override TakeTurn() and basically do what they used to
do in their Action property override, but set base.Action and then run
the base method, AbstractActor.TakeTurn to execute it.

The basically is to encapsulate what goes on during Actors turns and
will make it easier to implement AI later on.