Actor: Made events for AI (and other) use
Actor interface has two new events: ActorSighted and ActorLostSight
(this should probably be renamed to something better, but I can't think
of anything). When one Actor becomes with 5 (remove the hard code later)
grids of another, that ActorSighted event is fired.

When a Actor that is in sight becomes out of sight, the ActorLostSight
event is fired.