Scaffolding for "Pickup?" mini-dialog
What I'm calling a mini-dialog is what Nautilus and various other Gnome
programs use to relay information to the user without popping a dialog
in their face. In Nautilus, go to the Trash location to see what I'm
talking about.

I plan on using these for various "contextual" things. Right now, this
one will show itself when the Hero walks on top of an item. It will tell
them what it is and give them an option to pick it up. Or it can be
ignored.

The glade file has been slightly modified to give us a place to display
it.

The TurnLoop has been modified to give us an event when the TurnLoop
starts, so now we get notified when Round 0 begins.

New PickupMessageWidget (maybe rename later to PickupMiniDialog and
inherit from MiniDialog, I'll think about this more we I think of more
uses for mini-dialogs). All it does is show a Label when the user is
standing on an Item.