Fixed Direction.GetRandom() to be more random
Direction.GetRandom() used to create the Random object everytime it was
called. This meant if called in rapid succession, the Random objects
could be constructed uses the same seed and cause Direction.GetRandom()
to return the same Direction multiple times.

Now we have a private static field for the Random object and
Direction.GetRandom() will use it everytime.

Also added a public overload for GetRandom() that takes a Random object
to use to get a random Direction. If you use the Random objects
constructed with the same seed, Direction.GetRandom() will return the
same Direction(s) in the same order.

Unit tests included to test for randomness of Direction.GetRandom() and
to make sure Direction.GetRandom(Random) returns the same Direction each
time.