Generalized MonsterFactory and renamed it to ActorFactory
ActorFactory can create both Monster and Hero objects. Later on the
overload that creates a Hero will take a BirthOptions object parameter
to define the sort of Hero to create.

This also gets rid of the hack required where Hero had to force
MonsterFactory to raise its Created event for the Hero, too.

Maybe all this is the wrong approach. Instead of transparent events that
do all the work, maybe things should be slightly more explicit:

    Actor new_monster = ActorFactory.Create ("theresa");
    new_monster.Location =
        Game.Instance.Hero.Location.RandomLocationInRadius (2);
    Game.Instance.Add (new_monster);