New CheckExtensions extension methods
AssertContains<T>() extends IEnumerable<T> types. It throws an
ArgumentException if the IEnumerable<T> does not contain the item. Right
now it cheats a bit. It checks if self is an IList<T>, if it is it uses
its Contains() method. Otherwise, it throws a NotImplementedException.
I'll fix this later to work for all IEnumerable<T> types.

Also added AssertIsTrue() (used as a helper for AssertContains<T>()). It
simply throws an ArgumentException if its bool is false. Obviously,
we're missing AssertIsFalse() and unit tests for this method.