Fixed up Inventory and Item-related Actions
The Execute() methods of the DrinkAction, DropAction, and PickUpAction
now use the various CheckExtensions to check their arguments.

Same thing for Inventory's Add() and Remove() methods. Also we no longer
mess with the Item's Location when we add and remove it.

No unit tests yet. I'll fix up the old ones and commit them soon.