Two more bug fixes related to item pickup
The first bug prevented items from being picked up because of an
PickUpAction.Execute() assertion that made sure that the item's location
and the actor's location are the same. Location equality was actually
using reference equality, but that is no longer the case.

The second bug is in how the map would update itself. Map did not expect
an item's location to change to null. It does now, because we no longer
have ActorLocations.