WorldView now completely uses Cairo for drawing
Which means its incredibly slow.

== Bugs ==

Not all of the WorldView window is shown, when the ZoomSetting is
smaller than Normal. The bottom and right side are blank like the edge
of a Map.

Occasionally, other ZoomSettings than Normal show small amounts of space
between tiles. This occurs either horizontally, vertically, or both.

It's slow. With the window as small as it can be, it takes around 250ms
to draw the map. Before it was between 25 and 50ms. This is between 5-10
times slower.

This is definitely not the way forward, but it's was interesting.