Fixed attacking
Fixing this issue makes it obvious that Questar's AI system is broken.
The AI we're using is basically a state machine, but there is no way for
it to transition back when an actor is no longer adjacent. We could add
a ActorNoLongerAdjacent event, but this quickly gets ugly.

Consider this experimental with the likely action of the whole AI system
being backed out when a better solution presents itself.