Fixed crasher when an Actor dies
Comes with a test case that was failing before the fix. Problem was from
assuming when an Actor's location changes, that args.NewLocation won't
ever be null, which is the case when an Actor dies.

Revitalized MockActor for ActorFixture unit tests. AbstractActor no
longer calls Messages.Add() when an Actor is attacked or dies, because
it requires Gtk+, which won't work for test fixtures.

The message system should either be removed in favor of animations or
made independent of windowing system.