Map now initializes its grids in an effort to be more self-contained
Previously a Map had to be created from a IMapGenerator to ensure its
grids were properly initialized. Now it creates the grids and the
IMapGenerator just sets the grids' properties (like the Terrain
property).

Did a bit of refactoring inside Map to make it more readable. Added a
property to Grid, IsTerrainBlocking, so Map doesn't have to access a
Grid's properties which may be null.

Also added a unit test and fixed the Actor unit test to take advantage.