Refactoring WorldView drawing code
New proprety for Map: Map.Bounds, returns a Rectangle representing the
size of the map. This should probably be a simpler type like Size or
Dimensions.

WorldView.DrawGrid() no longer needs to check if the grid's invalid,
because we no longer attempt to draw grids outside of the map's bounds.