Implemented very basic field of view for the hero
For right now all grids inside an 11x11 rectangle around the Hero are
visible. This needs to be circular instead of a rectangle and also needs
to take into account any obstacles that can break line of sight.

Grids not in fov are dimmed in the WorldView and actors outside of fov
are not drawn at all.